+++ cli/src/commands/git/clone.rs
@@ -220,7 +220,7 @@ fn fetch_new_remote(
     let mut tx = workspace_command.start_transaction();
     let mut git_fetch = GitFetch::new(tx.repo_mut(), &git_settings)?;
     with_remote_git_callbacks(ui, |cb| {
-        git_fetch.fetch(remote_name, &[StringPattern::everything()], cb, depth)
+        git_fetch.fetch(remote_name, &[StringPattern::everything()], true, cb, depth)
     })?;
     let default_branch = git_fetch.get_default_branch(remote_name)?;
     let import_stats = git_fetch.import_refs()?;